Rename package to aeon namespace and remove rig model
* Rename package to overlap aeon implicit namespace
* Remove rig model

This model can come from a template but will always be
experiment-specific.
