+++ src/swc/aeon/rigs/base.py
